Precompiled Header
Using a precompiled header will greatly speed up the time it takes to
rebuild once we have all of the header files written.

The file for the precompiled header was added under the last commit as
"stdafx.h". When creating C++ classes, make sure to add the generated
header into "stdafx.h" and add the line:

#include "stdafx.h"

to both the header and cpp file. No other includes will be necessary as
the precompiled header will have all of the necessary includes inside of
it.